+++ config/apple-epf.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+
+    /*
+     * You can get those values from your Gracenote developer account
+     */
+    'user_id' => env('EPF_USER_ID'),
+    'password' => env('EPF_PASSWORD'),
+    
+];