+++ src/EPFServiceProvider.php
@@ -11,7 +11,11 @@ class EPFServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        //
+        if ($this->app->runningInConsole()) {
+            $this->publishes([
+                __DIR__.'/../config/apple-epf.php' => config_path('apple-epf.php'),
+            ], 'config');
+        }
     }
 
     /**
@@ -19,6 +23,6 @@ class EPFServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        //
+        $this->mergeConfigFrom(__DIR__.'/../config/apple-epf.php', 'apple-epf');
     }
 }