+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next (2017-03-18)
 - First draft of models for table, except `video_translation`, see readme's notes for explanation
+- Requires to have values in the `.env` file, for usage with `config/apple-epf.php`
 
 ## 0.0.1 (2017-03-18)
 - Base skeleton for the package, based on appstract/php-skeleton